Updated login authentication (moved to loggedin_homepage route - probably needs renaming at some point) and integrated the login system with the rest of the web app
