+++ private/app/templates/layout.html.twig
@@ -1,15 +1,19 @@
 <!DOCTYPE html>
 <html lang=”en">
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
 
-    <meta name="author" content="Matthew Trim, Christopher Sherlock, Sam Spencer-Mcfadden" />
-    <link rel="stylesheet" href="{{ css_path }}" type="text/css"/>
-    <script type="text/javascript" src="{{  js_path }}"></script>
-    <title>{% block title %}{% endblock %}</title>
-</head>
-<body>
-{% block banner %}{% endblock %}
-{% block content %}{% endblock %}
-</body>
+        <meta name="author" content="Matthew Trim, Christopher Sherlock, Sam Spencer-Mcfadden" />
+        <link rel="stylesheet" href="{{ css_path }}" type="text/css"/>
+        <script type="text/javascript" src="{{  js_path }}"></script>
+        <title>{% block title %}{% endblock %}</title>
+    </head>
+    <body>
+        {% block banner %}{% endblock %}
+        {% block content %}{% endblock %}
+        <div id="footer" class="cent">
+            <a href={{ path_for('login') }}>Home</a>
+            <a href={{ path_for('register') }}>Test</a>
+        </div>
+    </body>
 </html>
